f2fs-tools: build binaries in Mac
This patch modifies f2fs-tools to be built in mac.

+++ fsck/fsck.c
@@ -881,14 +881,14 @@ skip_blkcnt_fix:
 	}
 	if (ftype == F2FS_FT_SYMLINK && i_blocks && i_size == 0) {
 		DBG(1, "ino: 0x%x i_blocks: %lu with zero i_size",
-							nid, i_blocks);
+						nid, (unsigned long)i_blocks);
 		if (c.fix_on) {
 			u64 i_size = i_blocks * F2FS_BLKSIZE;
 
 			node_blk->i.i_size = cpu_to_le64(i_size);
 			need_fix = 1;
 			FIX_MSG("Symlink: recover 0x%x with i_size=%lu",
-							nid, i_size);
+						nid, (unsigned long)i_size);
 		}
 	}
